fix: concurrency map to store data + better executor queue handling + evaluation script
The frontend concurrency map is drastically reducing the RAM usage of the
frontend measurement type.
It also fixes the handler problem when 10M or more points were asked.
The handler is now working alongside the db insertion as we wanted and has a delay
of a small amount of seconds only.

The executor queue was filled even if tasks weren't needed, we have now a couple of task that
renew themselves with a fail safe mechanism that relaunch a task if the renew process or the task
has failed

I also add some documentation on the handler and an evaluation script that compares the frontend
measurement type with the Google's raw measurement type based on the operation/sec